# docs/openapi.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "The Source API",
    "version": "1.0.0",
    "description": "Read-only Project Gutenberg catalog, installed text, and random prose. English is the default for selection endpoints. GET and HEAD only. No authentication. Query parameters must be supplied once. Browse and random endpoints reject unknown query parameters. Random responses use Cache-Control: no-store. HTML UI routes are not API endpoints. This document uses JSON syntax, which is valid YAML 1.2."
  },
  "servers": [
    {
      "url": "/",
      "description": "Same server as this specification (default port 45068)"
    }
  ],
  "security": [],
  "paths": {
...
    "/api/v1/books": {
      "get": {
        "operationId": "listBooks",
        "summary": "Browse and search metadata",
        "description": "Text records only. Case-insensitive title order, then ID. Every whitespace-separated query word must occur as a substring in title, authors, subjects or bookshelves; no text-content search. Empty results return 200. Numeric offset cursors are not snapshots; a catalog refresh can change pages.",
        "parameters": [
          {
            "$ref": "#/components/parameters/language"
          },
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Metadata search; maximum 256 UTF-8 bytes.",
            "example": "austen"
          },
          {
            "name": "available",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ]
            },
            "description": "true selects installed texts; false selects absent texts; omission selects both. Blank invalid."
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 25,
              "minimum": 1,
              "maximum": 100
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string",
              "pattern": "^[0-9]+$"
            },
            "description": "Nonnegative offset returned as next_cursor. Omit for first page."
          }
        ],
        "responses": {
          "200": {
            "description": "Matching catalog records; original publication year is not populated here.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookPage"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Invalid"
          },
          "405": {
            "$ref": "#/components/responses/Method"
          }
        }
      },
      "head": {
        "operationId": "listBooksHead",
        "summary": "Browse and search metadata (headers only)",
        "description": "Same selection/validation as GET; no response body. Text records only. Case-insensitive title order, then ID. Every whitespace-separated query word must occur as a substring in title, authors, subjects or bookshelves; no text-content search. Empty results return 200. Numeric offset cursors are not snapshots; a catalog refresh can change pages.",
        "parameters": [
          {
            "$ref": "#/components/parameters/language"
          },
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Metadata search; maximum 256 UTF-8 bytes.",
            "example": "austen"
          },
          {
            "name": "available",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ]
            },
            "description": "true selects installed texts; false selects absent texts; omission selects both. Blank invalid."
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 25,
              "minimum": 1,
              "maximum": 100
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string",
              "pattern": "^[0-9]+$"
            },
            "description": "Nonnegative offset returned as next_cursor. Omit for first page."
          }
        ],
        "responses": {
          "200": {
            "description": "Matching catalog records; original publication year is not populated here."
          },
          "400": {
            "description": "400 invalid_query: invalid, unknown or repeated selection parameter; invalid range."
          },
          "405": {
            "description": "405 method_not_allowed. Allow: GET, HEAD."
          }
        }
      }
    },
    "/api/v1/books/{id}": {
      "get": {
        "operationId": "getBook",
        "summary": "Get book metadata",
        "description": "Retrieves the requested ID regardless of language or catalog type. Not subject to selection filters. Known publication year is enriched from the installed file header.",
        "parameters": [
          {
            "$ref": "#/components/parameters/id"
          }
        ],
        "responses": {
          "200": {
            "description": "Book record.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AvailableBook"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/Missing"
          },
          "405": {
            "$ref": "#/components/responses/Method"
          }
        }
      },
      "head": {
        "operationId": "getBookHead",
        "summary": "Get book metadata (headers only)",
        "description": "Same selection/validation as GET; no response body. Retrieves the requested ID regardless of language or catalog type. Not subject to selection filters. Known publication year is enriched from the installed file header.",
        "parameters": [
          {
            "$ref": "#/components/parameters/id"
          }
        ],
        "responses": {
          "200": {
            "description": "Book record."
          },
          "404": {
            "description": "404: book_not_found for unknown IDs; text_unavailable for absent text; no_matching_books for empty random selection; not_found for unknown API routes."
          },
          "405": {
            "description": "405 method_not_allowed. Allow: GET, HEAD."
          }
        }
      }
    },
...
    "/api/v1/books/random": {
      "get": {
        "operationId": "randomBook",
        "summary": "Pick a random installed book",
        "description": "Uniform among installed Text records matching language and optional publication-year bounds. Returns metadata, not book contents. Repeats are allowed. Years are indexed before serving requests and persisted across restarts. Selection uses sorted in-memory indexes with no request-time header I/O. Availability is a startup snapshot.",
        "parameters": [
          {
            "$ref": "#/components/parameters/language"
          },
          {
            "$ref": "#/components/parameters/year_from"
          },
          {
            "$ref": "#/components/parameters/year_to"
          }
        ],
        "responses": {
          "200": {
            "description": "Chosen book; available is true.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AvailableBook"
                }
              }
            },
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Invalid"
          },
          "404": {
            "$ref": "#/components/responses/Missing"
          },
          "405": {
            "$ref": "#/components/responses/Method"
          }
        }
      },
      "head": {
        "operationId": "randomBookHead",
        "summary": "Pick a random installed book (headers only)",
        "description": "Same selection/validation as GET; no response body. Uniform among installed Text records matching language and optional publication-year bounds. Returns metadata, not book contents. Repeats are allowed. Years are indexed before serving requests and persisted across restarts. Selection uses sorted in-memory indexes with no request-time header I/O. Availability is a startup snapshot.",
        "parameters": [
          {
            "$ref": "#/components/parameters/language"
          },
          {
            "$ref": "#/components/parameters/year_from"
          },
          {
            "$ref": "#/components/parameters/year_to"
          }
        ],
        "responses": {
          "200": {
            "description": "Chosen book; available is true.",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            }
          },
          "400": {
            "description": "400 invalid_query: invalid, unknown or repeated selection parameter; invalid range."
          },
          "404": {
            "description": "404: book_not_found for unknown IDs; text_unavailable for absent text; no_matching_books for empty random selection; not_found for unknown API routes."
          },
          "405": {
            "description": "405 method_not_allowed. Allow: GET, HEAD."
          }
        }
      }
    },
...
    "/api/v1/lyrics": {
      "get": {
        "operationId": "listLyrics",
        "summary": "Browse and search songs",
        "description": "Without q, ascending id order. With q, runs SQLite FTS5 full-text search over title, artist and lyrics body (each whitespace-separated word required, as a quoted term) and orders by relevance: title matches outweigh artist matches, which outweigh body matches, with view count breaking ties. Optional language and tag filters. Song lyrics bodies are omitted from listings. Empty results return 200. Numeric offset cursors are not snapshots.",
        "parameters": [
          {"$ref": "#/components/parameters/songLanguage"},
          {"$ref": "#/components/parameters/tag"},
          {"name": "q", "in": "query", "schema": {"type": "string"}, "description": "Full-text query; maximum 256 UTF-8 bytes.", "example": "concrete jungle"},
          {"name": "limit", "in": "query", "schema": {"type": "integer", "default": 25, "minimum": 1, "maximum": 100}},
          {"name": "cursor", "in": "query", "schema": {"type": "string", "pattern": "^[0-9]+$"}, "description": "Nonnegative offset returned as next_cursor. Omit for first page."}
        ],
        "responses": {
          "200": {"description": "Matching songs without lyrics bodies.", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/SongPage"}}}},
          "400": {"$ref": "#/components/responses/Invalid"},
          "404": {"$ref": "#/components/responses/LyricsMissing"},
          "405": {"$ref": "#/components/responses/Method"},
          "500": {"$ref": "#/components/responses/LyricsError"}
        }
      },
      "head": {
        "operationId": "listLyricsHead",
        "summary": "Browse and search songs (headers only)",
        "description": "Same selection/validation as GET; no response body.",
        "parameters": [
          {"$ref": "#/components/parameters/songLanguage"},
          {"$ref": "#/components/parameters/tag"},
          {"name": "q", "in": "query", "schema": {"type": "string"}, "description": "Full-text query; maximum 256 UTF-8 bytes."},
          {"name": "limit", "in": "query", "schema": {"type": "integer", "default": 25, "minimum": 1, "maximum": 100}},
          {"name": "cursor", "in": "query", "schema": {"type": "string", "pattern": "^[0-9]+$"}}
        ],
        "responses": {
          "200": {"description": "Matching songs without lyrics bodies."},
          "400": {"description": "400 invalid_query: invalid, unknown or repeated selection parameter."},
          "404": {"description": "404 lyrics_unavailable: no lyrics corpus installed."},
          "405": {"description": "405 method_not_allowed. Allow: GET, HEAD."},
          "500": {"description": "500 lyrics_error: the lyrics database could not be queried."}
        }
      }
    },
    "/api/v1/lyrics/{id}": {
      "get": {
        "operationId": "getSong",
        "summary": "Get song metadata",
        "description": "Retrieves one song by id. Lyrics body is not included; use the text endpoint.",
        "parameters": [{"$ref": "#/components/parameters/songId"}],
        "responses": {
          "200": {"description": "Song record without lyrics body.", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Song"}}}},
          "404": {"$ref": "#/components/responses/LyricsMissing"},
          "405": {"$ref": "#/components/responses/Method"},
          "500": {"$ref": "#/components/responses/LyricsError"}
        }
      },
      "head": {
        "operationId": "getSongHead",
        "summary": "Get song metadata (headers only)",
        "description": "Same selection/validation as GET; no response body.",
        "parameters": [{"$ref": "#/components/parameters/songId"}],
        "responses": {
          "200": {"description": "Song record without lyrics body."},
          "404": {"description": "404 song_not_found for unknown ids; lyrics_unavailable when no corpus is installed."},
          "405": {"description": "405 method_not_allowed. Allow: GET, HEAD."},
          "500": {"description": "500 lyrics_error: the lyrics database could not be queried."}
        }
      }
    },
...
    "/api/v1/lyrics/random": {
      "get": {
        "operationId": "randomSong",
        "summary": "Pick a random song",
        "description": "Random over songs matching the optional language, tag and view-count filters, independent of storage order (indexed random bucket, not a full scan). Returns metadata without the lyrics body. Repeats are allowed.",
        "parameters": [
          {"$ref": "#/components/parameters/songLanguage"},
          {"$ref": "#/components/parameters/tag"},
          {"$ref": "#/components/parameters/views_from"},
          {"$ref": "#/components/parameters/views_to"}
        ],
        "responses": {
          "200": {"description": "Chosen song without lyrics body.", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Song"}}}, "headers": {"Cache-Control": {"schema": {"type": "string", "const": "no-store"}}}},
          "400": {"$ref": "#/components/responses/Invalid"},
          "404": {"$ref": "#/components/responses/LyricsMissing"},
          "405": {"$ref": "#/components/responses/Method"},
          "500": {"$ref": "#/components/responses/LyricsError"}
        }
      },
      "head": {
        "operationId": "randomSongHead",
        "summary": "Pick a random song (headers only)",
        "description": "Same selection/validation as GET; no response body.",
        "parameters": [
          {"$ref": "#/components/parameters/songLanguage"},
          {"$ref": "#/components/parameters/tag"},
          {"$ref": "#/components/parameters/views_from"},
          {"$ref": "#/components/parameters/views_to"}
        ],
        "responses": {
          "200": {"description": "Chosen song without lyrics body.", "headers": {"Cache-Control": {"schema": {"type": "string", "const": "no-store"}}}},
          "400": {"description": "400 invalid_query: invalid, unknown or repeated selection parameter."},
          "404": {"description": "404 no_matching_songs when no song matches; lyrics_unavailable when no corpus is installed."},
          "405": {"description": "405 method_not_allowed. Allow: GET, HEAD."},
          "500": {"description": "500 lyrics_error: the lyrics database could not be queried."}
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "language": {
        "name": "language",
        "in": "query",
        "description": "Defaults to English, including multilingual records containing en. Explicit all removes the filter. Blank is invalid. Trimmed and case-insensitive.",
        "schema": {
          "type": "string",
          "minLength": 1,
          "default": "en"
        },
        "examples": {
          "english": {
            "value": "en"
          },
          "all": {
            "value": "all"
          },
          "french": {
            "value": "fr"
          }
        }
      },
      "year_from": {
        "name": "year_from",
        "in": "query",
        "description": "Inclusive original publication year lower bound. Unknown years excluded when either bound is set. Use 1901 for after 1900. Must not exceed year_to.",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 9999
        },
        "example": 1901
      },
      "year_to": {
        "name": "year_to",
        "in": "query",
        "description": "Inclusive original publication year upper bound; not Gutenberg release year.",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 9999
        },
        "example": 1950
      },
      "id": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "integer",
          "minimum": 1
        },
        "example": 1342
      },
      "songLanguage": {
        "name": "language",
        "in": "query",
        "description": "Optional language filter for songs. Unlike books, there is no English default: omission or all removes the filter. Trimmed and case-insensitive.",
        "schema": {"type": "string", "minLength": 1},
        "examples": {"english": {"value": "en"}, "all": {"value": "all"}}
      },
      "tag": {
        "name": "tag",
        "in": "query",
        "description": "Optional genre filter (e.g. rap, pop, rock, rb, country, misc). Trimmed and case-insensitive.",
        "schema": {"type": "string", "minLength": 1},
        "example": "rap"
      },
      "songId": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {"type": "integer", "minimum": 1},
        "example": 10
      },
      "views_from": {
        "name": "views_from",
        "in": "query",
        "description": "Inclusive lower bound on view count. Omit or 0 for no lower bound. Must not exceed views_to.",
        "schema": {"type": "integer", "minimum": 0},
        "example": 1000
      },
      "views_to": {
        "name": "views_to",
        "in": "query",
        "description": "Inclusive upper bound on view count. Omit or 0 for no upper bound.",
        "schema": {"type": "integer", "minimum": 0},
        "example": 100000
      }
    },
    "schemas": {
      "Book": {
        "type": "object",
        "required": [
          "id",
          "type",
          "issued",
          "title",
          "languages",
          "authors",
          "subjects",
          "locc",
          "bookshelves"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "minimum": 1,
            "example": 1342
          },
          "type": {
            "type": "string",
            "example": "Text"
          },
          "issued": {
            "type": "string",
            "description": "Gutenberg release date, NOT original publication year.",
            "example": "1998-06-01"
          },
          "title": {
            "type": "string",
            "example": "Pride and Prejudice"
          },
          "languages": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": [
              "en"
            ]
          },
          "authors": {
            "type": "string",
            "description": "Catalog display string; may contain multiple authors and dates.",
            "example": "Austen, Jane, 1775-1817"
          },
          "subjects": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "locc": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "bookshelves": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "original_publication_year": {
            "type": "integer",
            "minimum": 1,
            "maximum": 9999,
            "description": "Only an unambiguous four-digit year from the Original publication header. Omitted when unknown, and not populated in browse results."
          }
        }
      },
      "AvailableBook": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Book"
          },
          {
            "type": "object",
            "required": [
              "available"
            ],
            "properties": {
              "available": {
                "type": "boolean",
                "description": "Presence in the startup corpus index; restart after corpus changes."
              }
            }
          }
        ]
      },
      "BookPage": {
        "type": "object",
        "required": [
          "books",
          "total",
          "next_cursor"
        ],
        "properties": {
          "books": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AvailableBook"
            }
          },
          "total": {
            "type": "integer",
            "minimum": 0
          },
          "next_cursor": {
            "type": "string",
            "pattern": "^[0-9]*$",
            "description": "Empty means no next page. Otherwise pass unchanged with the same filters and limit."
          }
        }
      },
...
      "Song": {
        "type": "object",
        "required": ["id", "title", "artist", "views"],
        "properties": {
          "id": {"type": "integer", "minimum": 1, "example": 10},
          "title": {"type": "string", "example": "Killa Cam"},
          "artist": {"type": "string", "example": "Cam'ron"},
          "tag": {"type": "string", "description": "Genre tag. Omitted when unknown.", "example": "rap"},
          "language": {"type": "string", "description": "Language code. Omitted when unknown.", "example": "en"},
          "year": {"type": "integer", "description": "Release year; only plausible values are kept, junk is omitted.", "example": 2004},
          "views": {"type": "integer", "minimum": 0, "description": "Genius view count; useful as a popularity signal."},
          "features": {"type": "array", "items": {"type": "string"}, "description": "Featured artists, parsed from the source array. Omitted when none."},
          "lyrics": {"type": "string", "description": "Lyrics body. Present only where explicitly documented; omitted from listings, item metadata, random and excerpt responses."}
        }
      },
      "SongPage": {
        "type": "object",
        "required": ["songs", "total", "next_cursor"],
        "properties": {
          "songs": {"type": "array", "items": {"$ref": "#/components/schemas/Song"}},
          "total": {"type": "integer", "minimum": 0},
          "next_cursor": {"type": "string", "pattern": "^[0-9]*$", "description": "Empty means no next page. Otherwise pass unchanged with the same filters and limit."}
        }
      },
...
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              }
            }
          }
        }
      }
    },
    "responses": {
      "Invalid": {
        "description": "400 invalid_query: invalid, unknown or repeated selection parameter; invalid range.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Missing": {
        "description": "404: book_not_found for unknown IDs; text_unavailable for absent text; no_matching_books for empty random selection; not_found for unknown API routes.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Method": {
        "description": "405 method_not_allowed. Allow: GET, HEAD.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "LyricsMissing": {
        "description": "404: song_not_found for unknown song ids; no_matching_songs for empty random selection; lyrics_unavailable when no lyrics corpus is installed.",
        "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}
      },
      "LyricsError": {
        "description": "500 lyrics_error: the lyrics database could not be queried.",
        "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}
      },
...
    }
  }
}